@@ -1,29 +1,43 @@
 {
   "name": "server",
-  "version": "0.0.1",
-  "description": "Servidor",
-  "main": "server.ts",
+  "version": "1.0.0",
+  "description": "",
+  "main": "index.js",
   "scripts": {
-    "test": "npm run tsc && jasmine --config=./spec/support/jasmine.json",
-    "tsc": "tsc",
-    "start": "npm run tsc && node typeScript/server.js"
+    "start": "concurrently \"npm run build:watch\" \"npm run watch:data\"",
+    "build:watch": "env-cmd -f .env.${env} tsc-watch --onSuccess \"node dist/src/index.js\"",
+    "copy:json": "copyfiles -u 1 ./src/data/*.json ./dist/src",
+    "watch:data": "if [ ! -f \"./dist/src/data/products.json\" ]; then npm run copy:json; fi && onchange 'src/data/*.json' -- npm run copy:json",
+    "build": "env-cmd -f .env.${env} tsc && npm run copy:json",
+    "deploy": "env-cmd -f .env.${env} npm run build && vercel --prod",
+    "test": "env-cmd -f .env.${env} jest --verbose --coverage ./tests/**/*.spec.ts",
+    "prettier": "prettier --write 'src/**/*.{ts,js}'",
+    "lint": "eslint '**/*.ts'"
   },
-  "author": "Daniel Bastos",
-  "license": "MIT",
+  "keywords": [],
+  "author": "",
+  "license": "ISC",
   "dependencies": {
-    "@types/body-parser": "^1.16.5",
-    "@types/express": "^4.0.37",
-    "body-parser": "^1.18.2",
-    "express": "^4.15.4",
-    "request": "^2.34",
-    "nodemailer": "^6.1.1"
+    "cors": "^2.8.5",
+    "env-cmd": "^10.1.0",
+    "express": "^4.18.2",
+    "pino": "^8.10.0",
+    "typescript": "^4.9.5",
+    "uuid": "^9.0.0"
   },
   "devDependencies": {
-    "@types/jasmine": "^2.8.9",
-    "@types/request-promise": "^4.1.38",
-    "jasmine": "^2.8.0",
-    "jasmine-promises": "^0.4.1",
-    "request-promise": "^4.2.2",
-    "typescript": "^3.3.3333"
+    "@types/cors": "^2.8.13",
+    "@types/express": "^4.17.17",
+    "@types/jest": "^29.4.0",
+    "@types/prettier": "^2.7.2",
+    "@types/uuid": "^9.0.0",
+    "@typescript-eslint/parser": "^5.51.0",
+    "concurrently": "^7.6.0",
+    "copyfiles": "^2.4.1",
+    "eslint-plugin-no-unused-expressions": "^0.1.0",
+    "nodemon": "^2.0.20",
+    "onchange": "^7.1.0",
+    "prettier": "^2.8.4",
+    "tsc-watch": "^6.0.0"
   }
 }
