@@ -1,3 +1,4 @@
 export const environment = {
-  production: true
+  production: true,
+  apiUrl: process.env['API_URL'] || 'http://localhost:3000/api',
 };
